# backend/apiserver/api/model_trade.go
/*
 * cross-chain-hackathon api
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

import (
	"time"
)

type Trade struct {

	// generated by the API server
	Id int64 `json:"id"`

	EstateId string `json:"estateId"`

	Price int64 `json:"price"`

	Buyer string `json:"buyer,omitempty"`

	Seller string `json:"seller"`

	Type string `json:"type"`

	Requests []TradeRequest `json:"requests"`

	Canceled bool `json:"canceled"`

	UpdatedAt time.Time `json:"updatedAt"`
}